refactor to MapAggregate
