+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/FunctionRegistry.scala
@@ -235,7 +235,7 @@ object FunctionRegistry {
     expression[Remainder]("%"),
 
     // aggregate functions
-    expression[HistogramEndpoints]("histogram_endpoints"),
+    expression[MapAggregate]("map_aggregate"),
     expression[HyperLogLogPlusPlus]("approx_count_distinct"),
     expression[Average]("avg"),
     expression[Corr]("corr"),
